Fix for the preperation of the STABs binary.
The gist of the (now fixed) issue is that the Makefile didn't take into account
NOBITS for the bss section and assumed the size of the binary was the full size
of the in-memory image. As such, if a student had an uninitialized variable,
they could overwrite part of the tacked on STABs data if they initialize it at
run-time.